requirements.txt
fastapi
pydantic
flask
transformers>=4.37
torch
pandas
python-dotenv
praw
os
uvicorn
keybert
collections